+++ docker/options.go
@@ -1,4 +1,4 @@
-//   Copyright 2016 Wercker Holding BV
+//   Copyright © 2016,2018, Oracle and/or its affiliates.  All rights reserved.
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -21,6 +21,7 @@ import (
 	"time"
 
 	"github.com/wercker/wercker/util"
+	"golang.org/x/net/context"
 )
 
 // DockerOptions for our docker client
@@ -39,14 +40,14 @@ type Options struct {
 	CleanupImage      bool
 }
 
-func guessAndUpdateDockerOptions(opts *Options, e *util.Environment) {
+func guessAndUpdateDockerOptions(ctx context.Context, opts *Options, e *util.Environment) {
 	if opts.Host != "" {
 		return
 	}
 
 	logger := util.RootLogger().WithField("Logger", "docker")
 	// f := &util.Formatter{opts.GlobalOptions.ShowColors}
-	f := &util.Formatter{false}
+	f := &util.Formatter{ShowColors: false}
 
 	// Check the unix socket, default on linux
 	// This will fail instantly so don't bother with the goroutine
@@ -56,11 +57,11 @@ func guessAndUpdateDockerOptions(opts *Options, e *util.Environment) {
 
 	if _, err := os.Stat(unixSocket); err == nil {
 		unixSocket = fmt.Sprintf("unix://%s", unixSocket)
-		client, err := NewDockerClient(&Options{
+		client, err := NewOfficialDockerClient(&Options{
 			Host: unixSocket,
 		})
 		if err == nil {
-			_, err = client.Version()
+			_, err = client.ServerVersion(ctx)
 			if err == nil {
 				opts.Host = unixSocket
 				return
@@ -108,7 +109,7 @@ func guessAndUpdateDockerOptions(opts *Options, e *util.Environment) {
 }
 
 // NewDockerOptions constructor
-func NewOptions(c util.Settings, e *util.Environment) (*Options, error) {
+func NewOptions(ctx context.Context, c util.Settings, e *util.Environment) (*Options, error) {
 	dockerHost, _ := c.String("docker-host")
 	dockerTLSVerify, _ := c.String("docker-tls-verify")
 	dockerCertPath, _ := c.String("docker-cert-path")
@@ -140,6 +141,6 @@ func NewOptions(c util.Settings, e *util.Environment) (*Options, error) {
 	// We're going to try out a few settings and set DockerHost if
 	// one of them works, it they don't we'll get a nice error when
 	// requireDockerEndpoint triggers later on
-	guessAndUpdateDockerOptions(speculativeOptions, e)
+	guessAndUpdateDockerOptions(ctx, speculativeOptions, e)
 	return speculativeOptions, nil
 }